requirements.txt
cv_bridge==1.16.0
colorama==0.4.3
opencv_contrib_python==4.5.4.60
torchvision==0.7.0
pandas==0.25.3
torchinfo
rospy==1.15.14
scipy==1.3.3
numpy==1.17.4
matplotlib==3.1.2
torch==1.6.0
Pillow==9.1.1
PyYAML==6.0
scikit_learn==1.1.1
torchsummary==1.5.1
trimesh==3.16.1
pycollada==0.7.2
Rtree==1.0.1